refactor : 클래스 타입 변경
