+++ src/main/kotlin/controller/MainController.kt
@@ -1,15 +1,13 @@
 package controller
 
-import model.Repository
 import view.InOutView
-import java.util.Random
 
 class MainController {
     private val inOutController = InOutController(InOutView())
 
     fun runProgram() {
-        val repo = Repository(inOutController.getCarName(), Random())
-        val raceController = RaceController(inOutController.getTryCount(), repo, inOutController)
+        val cars = inOutController.getCarName()
+        val raceController = RaceController(inOutController.getTryCount(), cars, inOutController)
         raceController.fullRace()
         val winnerList = raceController.getFinalResult()
         inOutController.printFinalResult(winnerList)
