+++ src/main/kotlin/view/InOutView.kt
@@ -1,6 +1,6 @@
 package view
 
-import data.Car
+import model.Car
 
 enum class InOutConstant(val constant: String) {
     NAME("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분)."),
